+++ AWSCloudProject/src/main/java/com/example/awscloudproject/service/JWTService.java
@@ -1,49 +1,14 @@
 package com.example.awscloudproject.service;
 
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.Map;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
 
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class JWTService {
-Key key = Keys.secretKeyFor(SignatureAlgorithm.HS512);
+public interface JWTService {
+  String generateToken(String userName);
 
-  public String generateToken(String userName) {
-    Map<String, Object> claims = new HashMap<>();
-    return createToken(claims, userName);
-  }
+  String createToken(Map<String, Object> claims, String userName);
 
-  private String createToken(Map<String, Object> claims, String userName) {
-    return Jwts.builder()
-        .setClaims(claims)
-        .setSubject(userName)
-        .setIssuedAt(new Date(System.currentTimeMillis()))
-        .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60))
-        .signWith(key, SignatureAlgorithm.HS512)
-        .compact();
-  }
+  public String extractUserName(String token);
 
-  public String extractUserName(String token) {
-    return Jwts.parserBuilder()
-        .setSigningKey(key)
-        .build()
-        .parseClaimsJws(token)
-        .getBody()
-        .getSubject();
-  }
-
-  public boolean validateToken(String token, UserDetails userDetails) {
-    Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-    return true;
-  }
+  public boolean validateToken(String token, UserDetails userDetails);
 }
