+++ AWSCloudProject/src/main/java/com/example/awscloudproject/service/JWTServiceImpl.java
@@ -1,0 +1,49 @@
+package com.example.awscloudproject.service;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+import java.security.Key;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class JWTServiceImpl implements JWTService {
+  Key key = Keys.secretKeyFor(SignatureAlgorithm.HS512);
+
+  public String generateToken(String userName) {
+    Map<String, Object> claims = new HashMap<>();
+    return createToken(claims, userName);
+  }
+
+  public String createToken(Map<String, Object> claims, String userName) {
+    return Jwts.builder()
+        .setClaims(claims)
+        .setSubject(userName)
+        .setIssuedAt(new Date(System.currentTimeMillis()))
+        .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60))
+        .signWith(key, SignatureAlgorithm.HS512)
+        .compact();
+  }
+
+  public String extractUserName(String token) {
+    return Jwts.parserBuilder()
+        .setSigningKey(key)
+        .build()
+        .parseClaimsJws(token)
+        .getBody()
+        .getSubject();
+  }
+
+  public boolean validateToken(String token, UserDetails userDetails) {
+    Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
+    return true;
+  }
+}
